plugins/stashDisableAll/stashDisableAll/requirements.txt
stashapi-tools
requests
